Remove reflection from loops in database.clj.
Reflection can be very costly in both time and memory when it occurs
inside a loop. A few functions in database.clj performed reflective
calls over all segments of the classpath or all entries in all JAR files
on the classpath, the latter of which can easily number in the hundreds
of thousands. In the interest of performance, add type hints to avoid
reflection in these looped calls.