Respect current schema.
When checking if a table exists (table-exists?) the connection schema
should be respected otherwise it will return true for any matching table
name across all schemas.

If a user wishes to mantain a migration table in a different schema they
should be explicit rather than be dependent on search across all
namespaces.

Compatibility Note: This will break migration for anyone who was
dependent on the global search, they should be able to resolve this by
including the target namespace via :migration-table-name