Add modify-sql-fn config hook
Add a :modify-sql-fn key to the config map. If a function is provided
here, it will be called for each user-provided sql statement before it
is executed. If :modify-sql-fn is not provided, the identity function is
used so no change is applied.

This is intended for use with
http://2ndquadrant.com/en/resources/pglogical/ and similar systems,
where DDL statements need to be executed via an extension-provided
function.