Add thread cancellation point and return information from migrate-up*
When many migrations are being applied, and the underlying database
is slow, the migrations can take a long time.  If the thread running
the migrations is cancelled, it may not be interrupted until all the
migrations are complete.

To allow the thread to terminate earlier, this injects a
thread cancellation check into the `migrate-up*` function.

To allow callers to be able to detect partial migrations, the
result code for `:ignore` and `:failure` are also returned. Successful
migrations return nil.