Make it possible to pass in existing connection/data-source
This commit adds code to let library users pass a java.sql.Connection or a
javax.sql.DataSource instance in place of a Clojure map for the :db key in the
configuration map.

The connect* function will make a type check on the db value. If the value is a
DataSource it will get a connection from the data-source. If it is a Connection
it will simply use that connection. If not, it will default to creating its own
Connection as before.

TODO: Figure out how to test data-source based store creation.