Fix error if any migration lacked applied date (#237)
The sort on `:applied` in `completed-ids*` did not handle nil, meaning the presence of one or more migrations lacking an `applied` timestamp would cause it to break. Any sql migrations pre-0.9.1 will not have an `applied` value unless users have retroactively set it, and the schema allows NULLs, so migratus cannot assume `applied` is not nil.